+++ Player.h
@@ -17,7 +17,7 @@ enum class Upgrade;
 
 class Player : public Entity {
 	public :
-		Player(sf::Color color, sf::Vector2f startPos, int radius, float maxHealth, float speed, float healthUpgradeValue, float baseDamage, float damagesUpgradeValue, 
+		Player(sf::Color color, sf::Vector2f startPos, float radius, float maxHealth, float speed, float healthUpgradeValue, float baseDamage, float damagesUpgradeValue, 
 				float bulletFireRate, float fireRateUpgradeValue, int timesWeaponUpgraded);
 		void Init(GameManager* gameManager);
 		void Draw(RefsData data);
